+++ kazoo/client.py
@@ -560,7 +560,8 @@ class KazooClient(object):
             raise ConnectionLoss("No connection to server")
 
         peer = self._connection._socket.getpeername()
-        sock = self.handler.create_connection(peer, timeout=self._session_timeout / 1000.)
+        sock = self.handler.create_connection(
+            peer, timeout=self._session_timeout / 1000.0)
         sock.sendall(cmd)
         result = sock.recv(8192)
         sock.close()